Work nearly finished on article editor, a few minor adjustemtns remaining.
