+++ components/ArticleSharedComponents/ImageBlock.js
@@ -93,7 +93,7 @@ const FullWidthButton = styled.button`
 `;
 
 
-class ImageBlock extends Component {
+class EditorImageBlock extends Component {
     constructor(props) {
         super(props);
         this.blockKey = this.props.block.getKey();
@@ -187,152 +187,4 @@ const mapStateToProps = state => ({
     token: state.currentUser.token
 });
 //export default ImageBlock;
-export default connect(mapStateToProps)(ImageBlock);
-
-
-
-
-
-
-
-
-
-
-
-
-// class ImageBlock extends Component {
-//     constructor(props) {
-//         super(props);
-//         //console.log(this.props.foo)
-//         //console.log(this.props.block.getKey());
-//         //console.log(this.props.contentState);
-//         this.blockKey = this.props.block.getKey();
-//         this.imageFileInput = React.createRef();
-//         this.updateImages = this.updateImages.bind(this);
-//         this.checkForFile = this.checkForFile.bind(this);
-//         this.uploadImage = this.uploadImage.bind(this);
-//         this.makeFullWidth = this.makeFullWidth.bind(this);
-//         this.makeHalfWidth = this.makeHalfWidth.bind(this);
-//         this.focusBlock = this.focusBlock.bind(this);
-//     }
-
-//     componentDidMount() {
-//         if (typeof window !== null) {
-//             window.focusBlock = this.focusBlock;
-//         }
-//     }
-
-//     updateImages(imagesObject) {
-//         this.focusBlock();
-//         //console.log(window.getSelection());
-//         this.props.contentState.mergeEntityData(
-//             this.props.block.getEntityAt(0),
-//             {images: imagesObject}
-//         );
-//         //console.log(window.getSelection());
-//         setTimeout(() => {
-//             this.forceUpdate();
-//         }, 0);
-//     }
-
-//     makeFullWidth() {
-//         this.props.contentState.mergeEntityData(
-//             this.props.block.getEntityAt(0), 
-//             {fullWidth: true}
-//         );
-//         setTimeout(() => {
-//             this.forceUpdate();
-//         }, 0);
-//     }
-
-//     makeHalfWidth() {
-//         this.props.contentState.mergeEntityData(
-//             this.props.block.getEntityAt(0), 
-//             {fullWidth: false}
-//         );
-//         setTimeout(() => {
-//             this.forceUpdate();
-//         }, 0);
-//     }
-
-//     checkForFile(e) {
-//         e.preventDefault();
-//         const imageFile = this.imageFileInput.current.files[0];
-//         if (imageFile) {
-//             //this.image = imageFile;
-//             this.uploadImage(imageFile);
-//         }
-//     }
-
-//     async uploadImage(image) {
-//         const form = new FormData();
-//         form.append('image', image);
-//         const settings = {
-//             headers: {
-//                 'Authorization': `Bearer ${this.props.token}`
-//             },
-//             method: 'post',
-//             body: form
-//         };
-//         const imageUploadRequest = await fetch(`${rootApiUrl}/api/upload`, settings);
-//         const imageUploadResponse = await imageUploadRequest.json();
-//         const imagesObject = imageUploadResponse.images;
-//         this.updateImages(imagesObject)
-//     }
-
-//     focusBlock() {
-//         /*
-//             Essentially I need to identify what is currently selected. If it this block
-//             that is currently selected, then I don't need to do anything. If it isn't this
-//             block, I need to alter the current selection to this block.
-//         */
-//         const { block, blockProps } = this.props;
-//         const { getEditorState, setEditorState } = blockProps;
-//         const key = block.getKey();
-//         const editorState = getEditorState();
-//         const selectionState = editorState.getSelection();
-//         const contentState = editorState.getCurrentContent();
-//         const currentBlock = contentState.getBlockForKey(selectionState.getStartKey());
-//         console.log(selectionState);
-//         if (currentBlock.getKey() === key) {
-//            return;
-//         }
-//         const newSelection = new SelectionState({
-//             anchorKey: key,
-//             focusKey: key,
-//             anchorOffset: 0,
-//             focusOffset: 0,
-//             hasFocus: true
-//         });
-//         setEditorState(EditorState.forceSelection(editorState, newSelection));
-//     }
-
-//     render() {
-//         const currentEntity = this.props.contentState.getEntity(
-//             this.props.block.getEntityAt(0)
-//         );
-//         const { images, fullWidth } = currentEntity.getData();
-//         let src = '';
-//         if (images.hasOwnProperty('original')) {
-//             src = images.original.imageUrl;
-//         }
-//         return (
-//             <Figure fullWidth={fullWidth} onClick={this.focusBlock}>
-//                 <Image src={src} />
-//                 <div>
-//                     <FileInputLabel>
-//                         <FileInput 
-//                             type="file"
-//                             name="imageFileInput"
-//                             onChange={this.checkForFile}
-//                             innerRef={this.imageFileInput}
-//                         />
-//                         Upload a new image...
-//                     </FileInputLabel>
-//                 </div>
-//                 <HalfWidthButton onMouseDown={this.makeHalfWidth}>H</HalfWidthButton>
-//                 <FullWidthButton onMouseDown={this.makeFullWidth}>F</FullWidthButton>
-//             </Figure>
-//         );
-//     }
-// }
+export const ImageBlock = connect(mapStateToProps)(EditorImageBlock);
